@@ -18,23 +18,34 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  lite_rolling_switch: ^0.1.1
+  lite_rolling_switch: ^1.0.1
 
 
 
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.2
-  roslib: ^0.0.3
+  cupertino_icons: ^1.0.5
+  #roslib: ^0.0.3
+  bloc: ^8.1.1
+  flutter_bloc: ^8.1.2
+  freezed: ^2.3.2
+  freezed_annotation: ^2.2.0
+  roslibdart: ^0.0.1-dev+4
+  get: ^4.6.5
+  injectable: ^2.1.0
+  injectable_generator: ^2.1.4
+  get_it: ^7.2.0
+
 
 dev_dependencies:
+  build_runner: ^2.3.3
   flutter_test:
     sdk: flutter
 
